chore: simplify place page controller
