--- android/src/app/organicmaps/widget/placepage/Closable.java
@@ -1,6 +1,0 @@
-package app.organicmaps.widget.placepage;
-
-public interface Closable
-{
-  void closePlacePage();
-}
